globalserver/client_interface_clone/requirements.txt
google-pasta==0.1.7
grpcio==1.24.3
protobuf==3.10.0
numpy==1.17.3
bson==0.5.8
